Refactor BaseWorld into partial classes for clarity
Separated functionality of BaseWorld into partial classes:
- BaseWorld.Drawing for rendering methods
- BaseWorld.GameState for game state management
- BaseWorld.InputHandling for user input handling
- BaseWorld.Properties for properties and fields

Restored state management and sound functionality, improving
organization and maintainability of the codebase.
